Separate device and e2e perf tests
