+++ .github/workflows/perf-device-models-impl.yaml
@@ -238,8 +238,7 @@ jobs:
           commands: |
             pytest models/demos/yolov8x/tests/perf -m models_device_performance_bare_metal
             pytest models/demos/yolov8s/tests/perf -m models_device_performance_bare_metal
-            # Skip for #27008
-            # pytest models/demos/yolov8s_world/tests/perf -m models_device_performance_bare_metal
+            pytest models/demos/yolov8s_world/tests/perf -m models_device_performance_bare_metal
         if: ${{ !cancelled() && matrix.test-info.civ2-compatible && !matrix.test-info.job-set1 && !matrix.test-info.job-blackhole-set1 }}
         timeout-minutes: ${{ matrix.test-info.timeout }}
