+++ models/demos/yolov8s_world/tests/perf/test_e2e_performant.py
@@ -9,11 +9,10 @@ import torch
 from loguru import logger
 
 import ttnn
-from models.common.utility_functions import is_wormhole_b0, run_for_wormhole_b0
+from models.common.utility_functions import run_for_wormhole_b0
 from models.demos.utils.common_demo_utils import get_mesh_mappers
 from models.demos.yolov8s_world.common import YOLOV8SWORLD_L1_SMALL_SIZE
 from models.demos.yolov8s_world.runner.performant_runner import YOLOv8sWorldPerformantRunner
-from models.perf.device_perf_utils import check_device_perf, prep_device_perf_report, run_device_perf
 from models.perf.perf_utils import prep_perf_report
 
 
@@ -148,38 +147,4 @@ def test_perf_yolov8s_world_dp(
         weight_dtype,
         model_location_generator,
         resolution,
-    )
-
-
-@run_for_wormhole_b0()
-@pytest.mark.parametrize(
-    "batch_size, expected_perf",
-    [
-        [1, 115],
-    ],
-)
-@pytest.mark.models_device_performance_bare_metal
-def test_perf_device_yolov8s_world(batch_size, expected_perf):
-    subdir = "ttnn_yolov8s_world"
-    num_iterations = 1
-    margin = 0.03
-    expected_perf = expected_perf if is_wormhole_b0() else 0
-
-    command = f"pytest models/demos/yolov8s_world/tests/pcc/test_ttnn_yolov8s_world.py::test_yolo_model"
-    cols = ["DEVICE FW", "DEVICE KERNEL", "DEVICE BRISC KERNEL"]
-
-    inference_time_key = "AVG DEVICE KERNEL SAMPLES/S"
-    expected_perf_cols = {inference_time_key: expected_perf}
-
-    post_processed_results = run_device_perf(command, subdir, num_iterations, cols, batch_size)
-    expected_results = check_device_perf(post_processed_results, margin, expected_perf_cols, assert_on_fail=True)
-
-    logger.info(f"{expected_results}")
-
-    prep_device_perf_report(
-        model_name=f"ttnn_yolov8s_world{batch_size}",
-        batch_size=batch_size,
-        post_processed_results=post_processed_results,
-        expected_results=expected_results,
-        comments="",
     )
